+++ .env.test
@@ -68,4 +68,4 @@ ETHEREUM_BLAST_RPC_URL="https://eth-mainnet.public.blastapi.io"
 STARKNET_PRIVATE_KEY=0x76f2ccdb23f29bc7b69278e947c01c6160a31cf02c19d06d0f6e5ab1d768b86
 STARKNET_PUBLIC_KEY=0x3bb306a004034dba19e6cf7b161e7a4fef64bc1078419e8ad1876192f0b8cd1
 STARKNET_RPC_URL=http://localhost:9944
-MADARA_BINARY_PATH="/Users/bytezorvin/work/karnot/madara/"
+MADARA_BINARY_PATH="/path/to/madara"
