+++ .gitignore
@@ -5,7 +5,6 @@
 
 *.code-workspace
 .vscode
-db/
 
 
 lcov.info
