+++ crates/settlement-clients/starknet/Cargo.toml
@@ -11,7 +11,6 @@ c-kzg = { workspace = true }
 color-eyre = { workspace = true }
 crypto-bigint = { workspace = true }
 dotenvy.workspace = true
-env_logger = "0.11.5"
 lazy_static = { workspace = true }
 mockall = { workspace = true }
 reqwest = { workspace = true }
